+++ module/ulid/src/main/kotlin/io/github/siyual_park/ulid/converter/StringToULIDConverter.kt
@@ -1,4 +1,4 @@
-package io.github.siyual_park.ulid.converter.r2dbc
+package io.github.siyual_park.ulid.converter
 
 import io.github.siyual_park.ulid.ULID
 import org.springframework.core.convert.converter.Converter
@@ -7,8 +7,8 @@ import org.springframework.stereotype.Component
 
 @Component
 @ReadingConverter
-class BytesToULIDConverter : Converter<ByteArray, ULID> {
-    override fun convert(source: ByteArray): ULID {
-        return ULID.fromBytes(source)
+class StringToULIDConverter : Converter<String, ULID> {
+    override fun convert(source: String): ULID {
+        return ULID.fromString(source)
     }
 }
