+++ module/ulid/src/main/kotlin/io/github/siyual_park/ulid/jackson/ULIDDeserializer.kt
@@ -1,4 +1,4 @@
-package io.github.siyual_park.ulid.converter.jackson
+package io.github.siyual_park.ulid.jackson
 
 import com.fasterxml.jackson.core.JsonParser
 import com.fasterxml.jackson.databind.DeserializationContext
