+++ module/ulid/src/main/kotlin/io/github/siyual_park/ulid/jackson/ULIDModule.kt
@@ -1,4 +1,4 @@
-package io.github.siyual_park.ulid.converter.jackson
+package io.github.siyual_park.ulid.jackson
 
 import com.fasterxml.jackson.databind.module.SimpleModule
 import io.github.siyual_park.ulid.ULID
